Show the true number of available seats in the course api
Added a new course resource, to handle overrides to toArray().
Now exposing the REAL number of free seats, taking reserved seats into
account.
If the order reserves all seats or tries to reserve seats, but 0 are
available, we issue the sync event for the course.
Moved getAvailableSeats into the course model.