You should be able to close an order, when all information is correct
Using the new route, an order can now be closed and marked as ready for
maconomy sync.
Added a method in OrderService to close the order.
Changed the values in the STATE_CONFIRMED from 2 to 3, on the Order
model.
Added a new STATE_CLOSED on the Order model, with value 2.
Updated api documentation.