An order can have multiple courses
Added a migration to add a new course_order table, to act as the
middleman between courses and orders.
Added a migration to migrate the data, so the course_id is no longer
on the order, but instead in the course_order table.
Made changes to the Course and Order models, to have a belongsToMany
relationship between those two