You can now get data from a single order
Changes for handling multiple courses in the Resources/Order model.
Now hiding the pivot table in the course.
Changed the select in getAvailableSeats, since an order no longer has a
course_id attached. This is now located in the course_order table
instead, hence the new left join.
Added company relation to the order model.
Added methods to get waitinglist and state as texts, so api users has an
easy way of seeing what "1 or 2" means.
Added a comment in the route file.