Handle errors from maconomy, since it doesn't work atm
Added exception handling for getSeatsAvailable(), since maconomy is
having a lot of issues atm. This fix simply logs the error to sentry
and moves on.

Removed the need to setting a logger on the maconomy client.

Added a migration to allow empty values on company name.
Added a migration to allow empty values on participant's maconomy_id,
since this is first set, when we get a response from maconomy service.