When reserving seats on an order, you cannot go beyond available seats
Before we just added the entire order, to waiting list.  This should no
longer be done, since they wanted them to know, that if there were
only 2 seats available, it would be better to just book those and then
create a new order with people the should be sent on the waiting list.
I simply readded the check to see if there are enough seats available
on the courses.

ILI-422