Make the booker and participant emails work
Changes for making the two email designs work, using a single layout
file, which is then extending the the two scripts, to ensure all
data is shown correctly.

Added a new intro_block section, to be able to provide "more" data
in the intro section instead of just a text.

ILI-279