You should be able to sync a participant on an order
Changes to the data fields for each participant, by adding a "Field"
after each field. Also changed what fields we are sending, based on
ticket ILI-465.

Added a new exception class to throw, if the participant data is
invalid.

Updated the phpdoc with throws on SyncOrder job.

The "last sync date" is now also saved on the order, just to make it
possible to debug at a later stage, what was sent when.

ILI-465