When clearing order reservations, tell wordpress to resync course
We are now looping through the courses on the order to clear, so we
can tell wordpress to update the course. This will correct the number of
seats available on the system.
Also added an ENV variable, to make it easier to change how long an
order reservation is available without code changes.