Trainers should be synced from wordpress
Added migrations and models for handling trainers and the integration
to a course.
Added a new controller, for handling syncs from wordpress to our api.
Started work on the mail handler.
Added a job, for sending out the participant list to the trainers.
Added "rough" email templates.
Updated api doc, with new trainer routes.
The Mail/Helper can now get the mailer object, with or without the
BCC mail added.
Updated course model, to have trainer reference.
Updated coursecontroller, so we can save trainers on the course.

ILI-230