You should be able to signup for a newsletter
Added a new controller to handle receiving the data to signup for the
newsletter. This is sent from the frontend, using their proxy.
The newsletter signup service, is injected using a service provider.
Added a new newsletter/signup route.
Added a newsletter service to send the data to clickdimensions.
Updated .env.example with the new required newsletter url.
Updated the api doc.

ILI-185