Add cron job, to send out participant lists to the trainers every day
Added a cron job, that sends out reminders for a course, to the trainers
on the course. The reminders are sent 1 day and 5 days before the course
is scheduled to start.
The reminders are sent at 7am in the morning, and we set a bool on the
course, that the reminders have been sent for it.
This way, we can always "unset" the reminder we want to set, and send it
out again if needed.

ILI-230