We should update wordpress more, when reserving seats
When an order takes all seats, we send an update to wordpress, so their
cache can be busted for the given course. This is done to allow their
interface to be updated a lot more, to give more accurate numbers.
Added a cast to int, since we sometimes got null out.

ILI-660