Booking to waitinglists, should have a different wording in emails
Changed the wording for the booker (only danish atm), so it now includes
"venteliste" as well :)
Added a new mail text constant, so we can use that in the email, when
sending mails to the participants. The text itself, comes from WP, so
we are just showing what we got from them.

ILI-640