Participant maconomy id, should be unique per course
Added a new table to handle storing multiple maconomy ids on the same
participant on the same order.
This means, that if a participant is signing up for more than one course
on a single order, the same participant should receive multiple maconomy
ids.
This is now stored in a single table, and used around the system.
To make the participant unique for a given course, we are now sending
courseId.participantId to maconomy (notice the period) instead of only
the participantId. This makes the same participant unique, per course on
a single order.

Added a new participant jsonresource, to choose what i am sending in the
feed.
Changes to the order adapter, to handle using the new participant
maconomy table.

ILI-717