Changes to handle sending out participant lists
Added a api route, sending out participantlist emails (with a given type)
This does NOT update the course, so you can resend again and again.
Think of it as a way of testing and doing support for a given course.
Changes to email layout.

NOTE: cronjob is NOT enabled.

ILI-230