Send participant list to trainers
We now fetch the list of participants from maconomy, instead of relying
on the data we have. This is done because implement can add bookings
directly in maconomy as well, therefore the trainers need the full list
of participants, and not only from the data we have.

Added a whitelist check, so we will avoid sending emails on dev or local
when testing, to "real people"!

Moved Order specific participant and response models, into an order
sub folder (namespace) in maconomy, to make it easier to separate from
other models.

Added a new participant model in maconomy, so we easily can get data
from the server.

Added a new method, to get the list of participants from the server.
While parsing the participant data from the server, we are not getting
enough data, so i have added some empty data for now.

ILI-531