You should be able to set the "before course" header in participant mail
Added a new type, so we can save the before course header and use this
in the emails.
If the before course header is empty, we use the old texts.
The empty check and language check is now moved into the Mailable
instead of having the logic in the view file.

ILI-731