When using a course day, the title should be on the course language
When showing a course day, we should use the course language (e.g.
English), to display the course day title.
The language specific course day title, should only be used here:
* yellow sticky box
* at checkout (selects and green box)
* receipt
* mailreceipt

When initializing the Resources/Course, we now use the course language,
to set what language the title should be displayed in. This can be
changed by calling setLanugauage on the Resources/Course model, if
needed.

When getting an order response, we no longer change the language of the
course day title, based on the order language.

ILI-741