Move calendar generation into own module
Added a new service class for handling generating the calendar for a
given order.
Added a method for generating the url for a given order, into the new
service class as well, called getLink(). This method uses the env
variable called SERVER_URL. If you wish to override this, you can do
that by calling setBaseUrl() on the service class as well.
Added location to the calendar's events', if a location has been added
to the course.

ILI-472