Show participants on waiting list in the email as well
Participants on waiting list, are now "saved" as well, when fetched from
maconomy.
We now sort the participants in two arrays: one with the signed up
participants and one with the participants on the waiting list.
Added a method for the Maconomy/Participant model, to check if a
participant is on the waiting list.
Changed the email layout, so it now lists the participants that are on
the waiting list, in a separate block.

ILI-230