Older orders should be deleted, when the courses are over
According to GDPR, we cannot store the orders and customer information
forever.
I add this cleanup job to remove old orders, where all the courses on
that order is over - and has been for 3 months.

Added a new "on delete" constraint to course_order table, to delete
all data in the system related to an order.

Added the job to the kernel file, and it's set to run everyday at
4am.

ILI-741