Participant email header and image is not working
Seems the email header was never used, so i added the possibility to set
this on the participant email. Default header will be shown, if no
header is set on any of the chosen courses.

The header image was hardcoded in the layout file. This now needs to be
set from every mail template.

Added a method in OrderParticipant to set the email header from the
given course type texts (if present).

Intro text was set twice in OrderParticipant.

ILI-720