Order booker emails should change the rule for subject course naming
Added a new getTitle method to CourseType, so we can modify what to
return, when getting the title for an education.
(default language is "blank" - meaning "da").
This is now used, in Helper::getTitle(Order), so we can get an
education's title, based on the language of the Order (meaning the
language of the site, the order was made on)-

ILI-792